feat(Fieldset)!: remove top-level sub-component(s)
