--- src/components/FieldsetItems/FieldsetItems.module.css
@@ -1,10 +1,0 @@
-/*------------------------------------*\
-    # FIELDSET ITEMS
-\*------------------------------------*/
-
-/**
- * The contents of the fieldset. Spaces them apart.
- */
-.fieldset-items > :not(:last-child) {
-  margin-bottom: var(--eds-size-1-and-half);
-}
