--- src/components/FieldsetItems/index.ts
@@ -1,2 +1,0 @@
-export { FieldsetItems as default } from './FieldsetItems';
-export type { FieldsetItemsProps } from './FieldsetItems';
