--- src/components/FieldsetLegend/index.ts
@@ -1,2 +1,0 @@
-export { FieldsetLegend as default } from './FieldsetLegend';
-export type { FieldsetLegendProps } from './FieldsetLegend';
